+++ DPLocalization/DPLocalization.h
@@ -14,3 +14,4 @@
 #import "DPAutolocalizationProxy.h"
 #import "Plural+DPLocalization.h"
 #import "DPFormattedValue.h"
+#import "DPLocalizationBundle.h"