+++ Example/LocalizationDemo/SecondViewController.m
@@ -26,7 +26,16 @@
     DPFormattedValue *numberValue = [DPFormattedValue formattedValueWithValue:@(self.stepper.value) formatter:self.numberFormatter];
     [self.countLabel updateAutolocalizationArguments:@[numberValue]];
 
-    self.bundleSelector.selectedSegmentIndex = ([[DPLocalizationManager currentManager].defaultBundle isEqual:[NSBundle mainBundle]]) ? 0 : 1;
+
+    if ([[DPLocalizationManager currentManager].defaultBundle isEqual:[NSBundle mainBundle]] == YES) {
+        self.bundleSelector.selectedSegmentIndex = 0;
+    }
+    else if ([[DPLocalizationManager currentManager].defaultBundle isKindOfClass:[DPLocalizationBundle class]] == YES) {
+        self.bundleSelector.selectedSegmentIndex = 2;
+    }
+    else {
+        self.bundleSelector.selectedSegmentIndex = 1;
+    }
 }
 
 - (IBAction)stepperValueChanged:(id)sender {
@@ -46,13 +55,35 @@
             break;
         }
 
+        case 2: {
+            DPLocalizationBundle *bundle = [DPLocalizationBundle defaultBundle];
+            NSDictionary *table = [bundle stringsTableWithName:nil language:@"en"];
+
+            NSError *error = nil;
+            table = @{
+                      @"TITLE": @"Runtime",
+                      @"LABEL_TEXT": @"Runtime - <traits=ibus size=12>{%1$@}\n%2$@\n<names=\"Courier-BoldOblique\" color=12,56,189>{%3$@}",
+                      @"TEXTVIEW_TEXT": @"Runtime\n\nUITextView text. <link=\"https://github.com/nullic/DPLocalizationManager\" traits=b>{link to DPLocalizationManager} ",
+                      };
+            [bundle setStringsTable:table withName:nil language:@"en" error:&error];
+
+            table = @{
+                      @"TITLE": @"Русский",
+                      @"LABEL_TEXT": @"Русский - %1$@\n%2$@\n%3$@",
+                      @"TEXTVIEW_TEXT": @"Русский\n\n<link=\"https://github.com/nullic/DPLocalizationManager\" traits=b>{смотри сюда} ",
+                      };
+            [bundle setStringsTable:table withName:nil language:@"ru" error:&error];
+
+            [DPLocalizationManager currentManager].defaultBundle = bundle;
+        }
+
         default:
             break;
     }
 
     NSLog(@"Preffered language: %@", [[DPLocalizationManager currentManager].defaultBundle preferredLanguage]);
     NSLog(@"Selected language: %@", dp_get_current_language());
-    NSLog(@"Supported language: %@", [[DPLocalizationManager currentManager].defaultBundle supportedLanguages]);
+    NSLog(@"Supported language: %@", [[DPLocalizationManager currentManager].defaultBundle localizations]);
 }
 
 @end